+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 -   Support on Windows operating systems
+-   Add PHP extension `calendar` as requirement
 
 
 ### Fixed